refactor: createPathPositions 메서드 ChessBoardPosition 클래스로 이동
